requirements.txt
streamlit
pandas
google-generativeai
plotly
numpy
PyGithub
reportlab